Aggregated position updating + updated fixture
The algorithms for flock centring, collision avoidance and velocity
matching repeatedly updated the positions. Instead these changes were
aggregated together using a local variable and then finally updated the
position at the end of the loop. These changes require the fixtures to
be updated for the regression test to pass.