Created shift_boid function with a unit test
Split the algorithms for flock centring, collision avoidance and
velocity matching into shift_boid, a separate function. This required
moving the parameters flock_attraction, avoid_radius, flock_radius and
speed_matching to be global variables. A unit test for this the new
shift_boid function was created using the default values for the global
values.