Added frame count logic to close plot window
The animation was set to run for 50 frames, however this would run
infinitely as there was not stop at the last frame. The generated
animate function now closes upon on reaching frame 50. The unit tests
have been updated to check this and the allowed maximum frame was
increased to 150 frames.